+++ backend/http/preview.go
@@ -97,7 +97,7 @@ func rawFileHandler(w http.ResponseWriter, r *http.Request, file iteminfo.Extend
 
 func previewHelperFunc(w http.ResponseWriter, r *http.Request, d *requestContext) (int, error) {
 	previewSize := r.URL.Query().Get("size")
-	if previewSize != "small" {
+	if !(previewSize == "small" || previewSize == "original") {
 		previewSize = "large"
 	}
 	if d.fileInfo.Type == "directory" {