+++ frontend/src/views/files/Preview.vue
@@ -1,7 +1,7 @@
 <template>
   <div id="previewer" @mousemove="toggleNavigation" @touchstart="toggleNavigation">
     <div class="preview">
-      <ExtendedImage v-if="previewType == 'image'" :src="raw"> </ExtendedImage>
+      <ExtendedImage v-if="previewType == 'image' || pdfConvertable" :src="raw"> </ExtendedImage>
       <audio
         v-else-if="previewType == 'audio'"
         ref="player"
@@ -111,6 +111,29 @@ export default {
     };
   },
   computed: {
+    pdfConvertable() {
+      const ext = "." + state.req.name.split(".").pop().toLowerCase(); // Ensure lowercase and dot
+      const pdfConvertCompatibleFileExtensions = {
+        ".pdf": true,
+        ".xps": true,
+        ".epub": true,
+        ".mobi": true,
+        ".fb2": true,
+        ".cbz": true,
+        ".svg": true,
+        ".txt": true,
+        ".doc": true,
+        ".docx": true,
+        ".ppt": true,
+        ".pptx": true,
+        ".xls": true,
+        ".xlsx": true,
+        ".hwp": true,
+        ".hwpx": true,
+        ".md": true,
+      };
+      return (!!pdfConvertCompatibleFileExtensions[ext]);
+    },
     sidebarShowing() {
       return getters.isSidebarVisible();
     },
@@ -118,6 +141,9 @@ export default {
       return getters.previewType();
     },
     raw() {
+      if (this.pdfConvertable) {
+        return filesApi.getPreviewURL(state.req.source, state.req.path, state.req.modified, "original")
+      }
       return filesApi.getDownloadURL(state.req.source, state.req.path, true);
     },
     isDarkMode() {